+++ intelmq/tests/bots/collectors/stomp/test_collector.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """
-Testing n6 stomp collector
+Testing stomp collector
 """
 import os
 
 if os.environ.get('INTELMQ_TEST_EXOTIC'):
-    import intelmq.bots.collectors.n6.collector_stomp
+    import intelmq.bots.collectors.stomp.collector